machine_learning/requirements.txt
clickhouse-driver
joblib
pandas
python-dotenv
riotwatcher
scikit-learn
tqdm
xgboost
